extract field logic to useField() hook
