+++ src/components/GameOfLife/GameOfLife.tsx
@@ -1,12 +1,6 @@
 import React, { useState, useEffect, useCallback } from 'react'
-import {
-	getInitialState,
-	invertOneCell,
-	calculateNextField,
-	// eslint-disable-next-line no-unused-vars
-	Field,
-} from './engine'
 import { GameField } from './components/GameField/GameField'
+import { useField } from './useField'
 
 interface GameOfLifeProps {
 	rowCount: number
@@ -19,31 +13,17 @@ export const GameOfLife: React.FC<GameOfLifeProps> = ({
 	colCount,
 	cellSize,
 }) => {
-	const [field, setField] = useState<Field>({
-		cells: [],
-		rowCount: 0,
-		colCount: 0,
-	})
+	const { field, clear, changeCell, step } = useField(
+		rowCount,
+		colCount,
+	)
 
 	const [autoplay, setAutoplay] = useState(false)
 
-	const playOneStep = useCallback(() => {
-		const newField = calculateNextField(field)
-		setField(newField)
-	}, [field])
-
-	const clear = useCallback((): void => {
-		setField(getInitialState(rowCount, colCount))
-	}, [rowCount, colCount])
-
 	const toggleAutoplay = useCallback(() => {
-		playOneStep()
+		step()
 		setAutoplay(!autoplay)
-	}, [autoplay, setAutoplay, playOneStep])
-
-	useEffect(() => {
-		clear()
-	}, [clear])
+	}, [autoplay, setAutoplay, step])
 
 	useEffect(() => {
 		if (!autoplay) {
@@ -51,23 +31,18 @@ export const GameOfLife: React.FC<GameOfLifeProps> = ({
 		}
 
 		const timeout = setTimeout(() => {
-			playOneStep()
+			step()
 		}, 300)
 		return () => {
 			clearTimeout(timeout)
 		}
-	}, [autoplay, playOneStep])
-
-	const changeCell = (row: number, col: number) => {
-		const newField = invertOneCell(field, row, col)
-		setField(newField)
-	}
+	}, [autoplay, step])
 
 	return (
 		<>
 			<GameField field={field} cellSize={cellSize} clickCell={changeCell} />
 			<button onClick={clear}>Clear</button>
-			<button onClick={playOneStep}>Step</button>
+			<button onClick={step}>Step</button>
 			<button onClick={toggleAutoplay}>{autoplay ? 'Stop' : 'Start'}</button>
 		</>
 	)
